Major rewrite of SQL store: less lines of code, less bugs, more features.
Especially, custom code for n-ary properties was replaced by a new generic support for "container" properties which also encompass "internal objects" as proposed by some extensions.

Also, the parts of the DB layout that are specific to the types of data that is stored is now is no longer hard-coded but governed by global data records, making way for future extensions and restrictions of the DB setup without changing the code. The store no longer needs to know all datatype classes to work properly, so datatype extensions now are better supported.

Various @todos remain, but all essential features are working as before.