Fix Special:Ask htmlspecialchars line 465
Special:Ask returns with a message htmlspecialchars() expects
parameter 1 to be string, array given. This is caused by line 465 where
$param->getValue() (IParam) returns with an array and not with a string.

In case of param sorting, values are returned as array and not as string
which causes htmlspecialchars to go south.

Tried to use  ->getOriginalValue() but it did not brougth any results
back therefor we stick with $param->getValue() to fetch the values

Change-Id: Ibc94b1471bb52745924310afdc2c65c39a0b8ebe