removing the Factory class for DIHandlers
This commit removes the factory class for DIHandlers
instead the handlers are created by the store3 and cached as well
to ensure efficiency.
However, some places the Store3 itself is created newly and this
leads to inefficiency but this has to be fixed in a separate commit.

Patchset3: Modified as per comments, the
store object is no more created anywhere on demand and is passed

Patchset4: Type hinting and direct access

Change-Id: Ibf0bb1ff1e27362a58634f45a74f2de609e053aa