Add individual format help page link in Special:Ask
## Description
This change will enable individual formats to show a help link next to the
format selection box. This can improve user engagement and access to
help resources on the smw home page.

A link will be displayed with its language dependent name while the url
will point to help page from where the most recent resources are available
on [1].

Discussion about which convention the url should adhere can be found [1] but
in general the generated link will follow [2].

## Test case
http://www.semantic-mediawiki.org/wiki/Template:Help.Redirect
http://www.semantic-mediawiki.org/wiki/Help:Json
http://www.semantic-mediawiki.org/wiki/Help:Json/en
http://semantic-mediawiki.org/wiki/Help:Json/en-gb

## Note
[1] http://semantic-mediawiki.org/wiki/Thread:User_talk:Kghbln/Link_from_Special:Ask_to_individual_format_help_page

[2] http://www.semantic-mediawiki.org/wiki/Help:<format>/<language>

Change-Id: Iae36108d67ee31dda4029c77979466bc94ae8b94