Split $.smwTooltip into two methods (init and add)
## Description
Split $.smwTooltip into two methods (init and add). The add method can be
used on any element and handles the creation of the icon/content element
and also initializes the tooltip with a qtip2 instance.

## Usage

$( ... ).smwTooltip( 'add', {
	title: ...,
	type: ...,
	content:...
} );

I'd like to use a namespace for this such as smw.tooltip/
semanticMediawiki.Tooltip but since jquery and namespaces usage sucks
(in order to allow normal chaining behaviour ) I have to leave this for
and use fn.smw...

## Dependency
on change I88adbae3

Change-Id: Icc9c43e428189a029c589a0869541d3dc4d54251