Implement consequences of new Container handling
Container DIs have recently become virtual -- they are never stored as such but only used to encapsulate user input that leads to complex data. Various conseqeunces of this design have now been implemented:

* Record properties are declared as having di type wikipage
* SQLStore3 has no more special code for storing containers, no extra table either
* SMWRecordValue can be created from SMWDIWikiPage objects (but also continues ot work with SMWDIContainer objects)

The change should not lead to new problems in SQLStore2 either, although it still has more code for Containers (which is simply not used now). Containers remain relevant for querying the stores and they will continue to be supported there.

Change-Id: Ifd159d353a0fcd928567d5c2d09df4da5168d978