Caching mechanism for property table hashes
Property hashes are optionally read and cached when reading the smw_ids table anyway, but only if requested by the caller. The cache is kept very small (1) and is supposed to be used in another call of getPropertyTableHashes() that most likely follows very soon after.

This reduces the number of SELECT for getting property hashes significantly, since they can usually be combined with another SELECT immediately before them. In local tests, the number of reads to smw_ids went down by about 40% when refreshing the data.

Update: Modified protected helper method name for consistent naming.

Change-Id: I92c6ecf00811db52f920c73904402da7cc1cdaad