Store data about inline queries
This change makes #ask and #show create new information that is stored for the page they are on. Each query creates a subobject that is used to store data about it. Currently, query string, size, depth, and format are stored. This data can be queried as usual, e.g., to find out how many queries on the wiki use a particular format, or what are the largest queries.

Changes in details:
* New special properties are declared to store the data
* All language files are updated to include translations for these properties (most are English placeholders with TODOs)
* The new properties get fixed IDs in the store
* The serialisation of query strings fixed: it escaped < and > in subquery syntax. If the string is used in a context that requires escaping, it should be escaped there.
* #ask and #show functions updated to store the data

Patch set 2: amended to also declare the same fixed property IDs for SQLStore2.
Patch set 3: amended to remove obsolete serialization code that caused a problem with SQLStore2 (Note: calling serialize() on a SQLStore2 object leads to a fatal error, maybe because it has a reference to the MW DB handler? Interesting problem.)
Change-Id: Iccd5fa173d59f1e311a03d20171710148a6b3d94