Fixed concept cache creation for SQLStore3
Concept caching did not work with Store3 since it was disabled explicitly (easy to fix) and since concept DIs were not read correctly frmo the database. To address the latter, it is necessary to distinguish fields in the DI table from fields that should be retrieved when fetching the DI from the DB (because the concept table has two extra fields that are not relevant to DI creation). Now all DI handlers specify which information should be fetched from the DB to build their DIs, and since this is often just one row, the DI creation allows the use of a string instead of an array that contains a single string only. This safes memory, which is the main goal in this mechanism.

Concrete changes:
* All DIHandlers have a method getFetchFields() that works like getTableFields() but only for fields whose contents is needed to create the DI.
* The SQLStore3Reader uses these fields instead of the table fields to fetch semantic data
* The SQLStoreReader uses plain strings instead of arrays whenever only one value is fetched
* The DIHandlers slightly change their dataItemFromDBKeys() implementations to use plain strings instead of arrays if only one key is needed. Checks are added to make sure that we never mistake an array for a string or vice versa (no PHP warnings even on unusual circumstances).
* Minor documentation and style updates in the affected methods/files.

Change-Id: I3f8589b5bc8210d3866e07a9bfe8bdd513ec2688