Reworked and repaired Special:SearchByProperty
This special page did not sort smaller nearby values correctly (nearby values are displayed for numerical properties for which only very few or no exact results are found; try Modification date). Besides this bug, the code was a mess (still is to some extent, but it is better) and did quite  few things wrong. Works better now.

Main changes:
* Fix sorting of results (problem was due to a use of array_reverse)
* Fix query generation (queries were created by building strings and parsing them! now the query is just created as an object)
* Rename many variables and a method to make some sense
* Some added docs
* Separate parameter reading from execute(); new method now; better/clearer defaults
* Improved code clarity in various places
* Fixed wrong HTML generation in the case where a few exact matches are displayed in between inexact matches (the exact matches would be surrounded by as many <strong> tags as there are matches in the old code; works but makes no sense at all)

Change-Id: I390f514f4f4d76e51f0555123904c9ea32319ad1