Fixed Bug 33610
The special page code did not handle errors that can occur on property dataitem creation (e.g., if a special property ID from the database is no longer known). To fix this, it was not enough to catch this problem, because the number of results must be the same as requested (otherwise the special page will not offer navigation to the next page of results). The code was changed to return an error dataitem in this case. Several other changes were needed to appropriately process such errors in special pages using this code.

Note for review: some code in the two special page files has been moved to a new function. May show as a bigger change in the diff than it actually is.

Change-Id: Ibdebbc6b4149afc2a0548e67584580f1637d0f5d