Fix Bug 38811
Special:UnusedProperties was not checking if a property is special (pre-defined in SMW). This is fixed now, and thus the datatype information for unused special properties shows correctly and the standard tooltip is given. The formatting code for special properties is copied from Special:Properties, so the formatting will be the same as it is there.

The change depends on the recent updates done to fix Bug 33610.

Patch set 2: Also make a small change in the Proeprty datavalue to make sure that the CSS styles for highlighting special properties are really included whenever they are needed. Fixes the missing italics for special properties in all Special pages that list properties.

Change-Id: I0a49c1dae5e27ca12dd52e00fff6c3df1d0b006f