(Bug 31880) add column class (based on dataValue typeId)
Depending on the datavalue type a class attribute (_str, _dat, _num etc.)
is amended which can be used to adjust in-table formatting.

The class attribute is concatenated string of
smwtype + typeId (e.g smwtype_dat, smwtype_num)

dataValue typeID is language independant and is also available via the
SMWApi which means a similar approach can be applied (see [1]).

For example:

<td class="Has-event-end smwtype_dat">5 March 2013 05:00:00</td>

[1] https://gerrit.wikimedia.org/r/#/c/53544

Change-Id: I361c93b95240e0719929de12901ce63d7e5978f6