Add test case (query string with spaces in AskParserFunctionTest)
Bug 47073 mentioned some issues with spaces, therefore
add a test case (that wasn't there before) where a query string containes
spaces. Test works and returns an appropriate result.

This doesn't test the QueryPrecessor directly, only what returns from it
and is stored successively in QueryData.

It means that if their is an issue (Bug 47073) it is not related to
AskParserFunction class itself.

Change-Id: I972d99c2586026d33e9c4df2b56e5467c8728694