SMW\MessageFormatter class (deprecates smwfEncodeMessages())
Code coverage: 100%
CRAP: 26

smwfEncodeMessages() deprecated in 1.9 with a removal in 1.11

## Narative

Invoke message objects instead of string/text representation during
processing and only transform messages before the actual output.

Instead of doing  wfMessage( '...' )->inContentLanguage()->text()
just invoked a Message( '...' ) object and let the MessageFormatter
resolve the text/html representation based on the language used for
the output.

It means a class only has to return a Message object without the need
to know the language.

Change-Id: Id252311eedc9855c961f5b1dfd02ade1ed8d3079