Remove addPropertyValue logic from ParserData (further compartmentalize of ParserData)
Move addPropertyValue logic into the SemanticData addDataValue method.

The logic provided by addPropertyValue has no "real" contract to the
ParserData object (but due to legacy conformism was transferred from
SMWParseData into ParserData which dilutes the responsibility of the
ParserData object)

addPropertyValue should be removed from the ParserData but this
needs to be done in a follow up in order to keep tests stable.

Change-Id: I12b197e7b589561c03b8c44c07be5b763f82c307