Disable type enforcement for properties with names equal to type names.
Now if you try to use some property named like an SMW builtin type
(for example, "[[Telephone number::jqlqdjq]]"), its type will be
always forced to be Telephone number, i.e. you can't override it
with [[has type::]]. I think this is an unevident and undocumented
behaviour, so this patch disables it.

Change-Id: I4b4922c6d5df4d3455d667f98732e13bab403769