Improve entry point
This gets rid of the extra composer entry point by making the main
entry point usable by composer in such as way that certain tools
(such as php-coveralls) don't break.

The issue was caused by the double-load guard returning, rather then
causing the entry point code to not be executed.

Change-Id: I64c7c4717b734038f20b0cfa99f47e4ce174abf4