Improve DeleteSubjectJob when run as DeferredJob
I wasn't really satisfied on how the deletion job was forced to handle
historic data, where it solely depended on the Store's data fetch at the time
of execution and withit did influence as to when to schedule the JobQueue.

SMW 1.9 deployes a SemanticData serializer/deserializer which enables to
serialize SemanticData and can be used before the subject and its data
are removed.

Running [1] the serializer adds a diminutive overhead during the deletion
process which enables to delete a subject immediately without delay but enables
to schedule the refresh of associative entities as deferred jobs (with a lower
priority, scaled as background job [2]).

[1] 'smwgDeleteSubjectWithAssociatesRefresh' => true
[2] 'smwgDeleteSubjectAsDeferredJob' => true