Improve DeleteSubjectJob when run as DeferredJob
The deletion job was forced to neglect historic data (Store has no knowledge
of any revision data), where it solely depended on the Store's data fetch at
the time of execution which influenced as to when to schedule the update and
removal.

SMW 1.9 deployes a SemanticData serializer/deserializer which enables to
serialize SemanticData at any time and can be used before the subject
and its data are removed.

Running [1] the serializer adds a diminutive overhead during the deletion
process which enables to delete a subject immediately without delay but enables
to schedule the dispatch and refresh associative entities as deferred job [2].

[1] 'smwgDeleteSubjectWithAssociatesRefresh' => true
[2] 'smwgDeleteSubjectAsDeferredJob' => true