Run HHVM tests only [DNM]
Runs the TRAVIS hhvm environment setup on the SMW test suite. It is
probably to earlier to tell if MW or SMW can run on HHVM but in order to
uncover possible errors we use this branch exclusively.

## Note
No local changes should be made to this branch, only a merge from master
should be carried out.