Fix highlighting of results
SMWSearch did not supply search terms needed by MW to apply highlighting to the
search results. So MW's Searchhighlighter put an empty span between each and
every byte of the result snippets.

This patch will return a search term that will match on every line start of the
article text so that the highlighter will output the first few characters of
the first few lines of the articles contained in the search result

Those lines will be truncated after 75 characters or so, so this may not
really give the correct beginning of the article. It should be good enough as a
temporary solution.