Introduce SemanticDataCache
Parsing pages is an expensive endeavour and therefore liable to the amount
of time spent on each page during an update (re-parse).

Using the `SemanticDataCache` can reduce the burden and processing times significantly
by making a decision on whether a page needs a "real" update
(revId to identify the content status) or it is merely enough to re-fetch
the data from cache and provide the input.

The premis is that if the revision hasn't changed then the SemanticData
haven't changed either hence the data from cache can be used.

- If `rebuildData.php` is run with option `--no-cache` then no `SemanticDataCache` is used.
- In case the de-serialization returns with an error a re-parse of page is automatically triggered during the `UpdateJob` run.
- Currently `smwgCacheType` is used to indentify the cache type for building the `CompositeCache` for the `SemanticDataCache` object.

Sample using `rebuildData.php -s 1 -e 40 --runtime`:
- without-cache: Memory used: 24309648 (b: 8997048, a: 33306696) with a runtime of 82.89 sec (1.38 min)
- with-cache: Memory used: 14102776 (b: 8997048, a: 23099824) with a runtime of 36.71 sec