Strict type declaration for properties with import reference
A user-defined property with a reference to an imported property `Imported from`
could define a different datatype than that of the imported property.
This can introduce inconsistencies for type declarations in properties.

`TypeByImportPropertyAnnotator` will use the type definition if it finds
an import property reference `Imported from` and ignore any user specified
datatype.