Redirect property alias to base property
If `Property:Doi` is an alias for `Property:DOI` then this ensures that
it is always redirected to the base property.

This avoids confusion where `Property:DOI` specifies extras (`_IMPO`)
but when displayed as `Property:Doi` the information is not visible
due to `Property:Doi` being not `Property:DOI`.

`Geographic coordinate` (deprecated) forwards to `Geographic coordinates`

This only touches the article view for `SMW_NS_PROPERTY`.