EmbeddedQueryDependencyLinksStore to track/update pages with embedded queries
The biggest obstacle for #ask queries to display refreshed(new) results is
the `ParserCache` by which an article decides whether it should be re-parsed
or use the data available in cache.

`EmbeddedQueryDependencyLinksStore` tracks known embedded queries and its
dependencies. If a dependency (property, subject etc.) has been
altered then `ParserCachePurgeJob` will initiate an invalidation
of the `ParserCache` for those selected dependencies.

Once the `ParserCache` for an article has been invalidated the next view
to this article will schedule a re-parse including the embedded #ask queries.

Because performance is crucial during an update/edit/delete, only the
first 200 (`ParserCachePurgeJob::CHUNK_SIZE`) entities are updated
immediately (online) while the rest is added as batch job in order for
the job scheduler to run the task on a regular basis.

`ParserCachePurgeJob` is a light job and can (or better should) run at
least each 15 min to minimize the amount of possible backlogs.

- `EmbeddedQueryDependencyLinksStore` to track and store dependencies
  in the `smw_query_links` table
- `ParserCachePurgeJob` job the handled the invalidation of selected
  entities

Changes to queries are detected using the `SMW::SQLStore::AfterDataUpdateComplete`
hook together with the `CompositePropertyTableDiffIterator`.

refs 555, 1108