Add PropertyTableOutdatedReferenceDisposer, refs 498
When a property type is changed or data being deleted chances are that some
ID's remain in the ID table and to avoid a pile of garbage being collected in the ID_TABLE,
`PropertyTableOutdatedReferenceDisposer` will check on ID's whether they can
safely be removed or not (during the `rebuidData` run).

For the quick sweep run (without running Parser::parse) use
`rebuildData.php --shallow-update`.